Add recreate test for live migrate rollback not cleaning up dest allocs
We rollback from a failed live migration in two cases:

1. The pre_live_migration on the destination host fails. The
   _do_live_migration method calls _rollback_live_migration
   explicitly to cleanup the dest host.

2. The live migration in the virt driver fails, and the virt driver
   calls back to _rollback_live_migration in the ComputeManager.

Either way, the instance is not on the destination host, so just like
how we remove volume connections and unplug vifs from the destination
host, we need to also remove allocations for the destination node in
Placement.

This change adds a test to show that we are not currently cleaning up
allocations on the destination node when we rollback from a live
migration failure.

Change-Id: Icbd5d7ff41aa04f8f7934fdce9668762691a4a69
Related-Bug: #1715182
(cherry picked from commit 6c3a58ce8001c588fe225d60ec198c7141d74700)