Remove dest node allocations during live migration rollback
When a live migration fails or is cancelled, either during
pre_live_migration on the destination host or during the
actual live migration itself, we rollback from the failure/abort
by doing things like removing volume connections from the
destination host and re-setup the network on the source host.

As part of the rollback from a failed or cancelled live migration,
we also need to remove the allocations created in Placement for the
destination node, since the instance is not on the destination
node.

Change-Id: I7b70cf8d5233bd25bf865a1b2789640758493c2b
Closes-Bug: #1715182
Closes-Bug: #1714237