Fix race in delete allocation in ServerMovingTests
When an instance is deleted the allocation of that instance is freed
after the instance is destroyed in the db. The functional tests that
asserting resource allocation after instance delete are waiting for
the instance to disappear from the REST API. This made such tests
racy.

Fortunately the instance.delete.end notification is emitted after
the instance allocation is freed. This patch fixes the race in the
test by waiting for the instance.delete.end notification before
asserting that the allocation is freed properly.

Change-Id: I4ffd8eae73600eb2d4f6929ee6a7768adb80081d
Closes-Bug: #1719915