Set group_members when converting to legacy request spec
In Pike we converted the affinity filter code to use the RequestSpec
object instead of legacy dicts. The filter used to populate server
group info in the filter_properties and the conversion removed that.
However, in the conductor, we are still converting RequestSpec back
and forth between object and primitive, and there is a mismatch
between the keys being set/get in filter_properties. So during a
reschedule with a server group, we hit an exception
"'NoneType' object is not iterable" in the RequestSpec.from_primitives
method and the reschedule fails.

This adds 'group_members' to the _to_legacy_group_info method to set
the key.

Closes-Bug: #1719730

Change-Id: Icb418f2be575bb2ba82756fdeb67b24a28950746
(cherry picked from commit 9d6632a67d91fb3c5145c14ac38011e919d6d8c0)