Rename the 'recreate' param in rebuild_instance to 'evacuate'
When recreate=True, we're evacuating during the
rebuild_instance method.

This is not obvious given the method name, so this
change renames the variable at the top of the method
and uses the new 'evacuate' variable throughout the
rest (except for a couple of noted places).

We could rename the parameter with some RPC version
dancing, but that gets complicated.

Later patches can further clean the rest of the
internals up with the virt driver interfaces.

Change-Id: Ib7276dd23f7c7ba7bd7afe37147a8479feb4b333