Fix live migration grenade ceph setup
Grenade runs in singleconductor mode for queens
as of change:

  If4c82ca12fe7b8b1ca7cfd8181d24dbd8dad3baa

However, the nova configuration during ceph setup
was using NOVA_CPU_CONF which is /etc/nova/nova-cpu.conf,
which is not what we want when configuring nova.conf
for the compute service in singleconductor mode.

Devstack has similar logic for stuff like this, so
we just have to handle it here since we're in a special
snowflake.

The stable/queens systemd stuff is all removed too since
we run with systemd on the old pike side and can restart
services properly with systemd on the new queens side
during grenade live migration CI job runs.

Change-Id: Iccb8eb55a5cc2a3d08e7fd6e31c89b3b5f8d0c70
Closes-Bug: #1720191