Remove old compat code from servers ViewBuilder._get_metadata
We always have an Instance object here with the metadata
attribute pre-loaded, so we can drop the compat code.

The methods which call through to use this via show() are:

1. show a specific server's details
2. list servers with details
3. update a server
4. rebuild a server

And those all get the Instance with the metadata joined
when it's pulled out of the database, so we're also not
going to lazy-load anything here either.

Change-Id: Ic9fc07d7fc2d32e4521991918da0bf9dd5dc624c