Stop joining on system_metadata when listing instances
We don't have anything in the API that requires joining
on the instance_system_metadata table anymore when listing
instances.

Long ago we needed it because the instance.flavor was embedded
in instance.system_metadata, but that was migrated to instance_extra
starting back in Kilo, and you couldn't upgrade to Liberty
without migrating that data first.

So we no longer need this join, and this should make listing
servers faster, especially when you have thousands of instances
per cell.

Change-Id: I0f620be8fa45527afd2be41442520c6c4d0f67b1