rp: move RP._set_traits() to module scope
To be consistent with the other module-level functions that handle the
raw DB access and writes for resource provider-related data, this patch
moves the ResourceProvider._set_traits() @staticmethod to the module
scope. In addition to this move, we also remove the @remotable
decoration on the ResourceProvider.set_traits() instance method (since
no resource provider objects are sent over RPC) and we ensure that the
resource provider object itself resets changes after setting traits (since the
provider generation is incremented).

blueprint: de-orm-resource-providers

Change-Id: I7667a6a4cedfe8cc4abb0785530e0859fb0c8bcf