rp: remove CRUD operations on Inventory class
The nova.objects.resource_provider.Inventory class was only intended to
be a plain-old data (POD) class and never intended to be able to
create, update or delete as a single item. Rather, the ResourceProvider
class has a set_inventory() method that accepts an InventoryList object
and operates on a set of inventory records as an atomic unit.

This patch removes the old, unused code paths that added CRUD
operations for the Inventory class, turning Inventory into a POD class.

blueprint: de-orm-resource-providers

Change-Id: Id8b182b3074122fc0bdbe201960e174b8f8d52f0