doc: Don't confuse CPU pinning/NUMA as Hyper-V only
A recent'ish change in openstack-manuals [1] noted the new support for
NUMA topologies when using the Hyper-V driver. As part of this change, a
section was added that describes the configuration steps that were
necessary on Hyper-V hosts before booting instances. However, the way
this section is integrated gives the impression that NUMA support is a
Hyper-V only feature.

Correct this by moving this configuration step to the end of the
document and instead opting to link to it from higher in the doc.

[1] https://review.openstack.org/#/c/424102/

Change-Id: Ic8d9c1b35d52a26374763b5c0e4be79875814569
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>