api-ref: add note about rebuild not replacing volume-backed root disk
We've had several bugs about this over the years, and until we
actually decide to fix it (or not), we should point out the known
limitation that volume-backed root disks are not replaced during
a rebuild.

Like, if you have a volume-backed instance and rebuild with a new
image, the root disk is still the volume with the original image.

Change-Id: I145cab88f782e4b1e630cc432322bc8436413e71
Related-Bug: #1482040