Reproduce bug 1721652 in the functional test env
When bug 1709902 was fixed in I0df401a7c91f012fdb25cb0e6b344ca51de8c309
the fix assumed that when the _destroy_evacuated_instances() is called
during the init of the nova-compute service the resource tracker
already knows the compute node ids associated to the given compute
host. This is not true and therefore _destroy_evacuated_instances fails
with an exception and does not clean up the evacuated instance.

The solution for the original bug included a functional regression
test. However the functional test env is not capable of fully simulate
a nova-compute service restart. The service only stopped then started
again. This allows some in memory state of the compute service to
survive the simulated restart. This caused that the functional test
was not able to catch that the above described assumption is not
correct.

Related-Bug: #1721652
Related-Bug: #1709902
Change-Id: Icaf1bae8cb040b939f916a19ce026031ddb84af7