Deprecate allowed_direct_url_schemes and nova.image.download.modules
The only in-tree implementation of the nova.image.download.modules
extension point was nova.image.download.file which was removed in
Pike: I7687cc89545a7a8b295dd6535b4ccebc913a2e0e

At the time of that removal, there was an operators ML thread
asking if anyone was using this code, or the extension point,
and the answer was no (or no answer at all).

Since we have no in-tree implementation of this extension point
and the extension point itself is not maintained or documented,
and even the TransferBase base class was removed in Pike, we
should deprecate the extension point and the configuration option
associated with its use so that we can simplify our internal
glance API client code.

Note that the libvirt Rbd image backend which does support
direct_url / image locations configuration for fast clones is
still supported and unrelated to this code.

Change-Id: I13162ebc9050dd2c468d0f8b969b96409f60afa8