Fix sending legacy instance.update notification
The if_notifications_enabled decorator skips the execution of the
decorated function if the versioned notifications are not configured
to be emitted. The send_instance_update_notification() call was wrongly
decorated with this decorator as it not only sends versioned
notification but also send the legacy compute.instance.update
notification as well. This caused that the legacy instance.update
notification was not emitted when the notification_format config option
was set to unversioned.

As the _send_versioned_instance_update() call already has the decorator
the solution is simply to remove the decorator from the
send_instance_update_notification() call.

Closes-Bug: #1721843
Change-Id: I9904adeb3de60cff4e29f1ab3c95399bbe9ff2e7