Fix typo in test_prep_resize_errors_migration
The errors_out_migration_ctxt context manager
sets the Migration object status to 'error', not 'failed'
which is also used in some places in the code. There is no
technical difference between either, they mean essentially
the same thing to the end user, but if we're going to use
them let's be clear in code comments.

Change-Id: Id51f3e2524ae6ee25f8d382070aeed9d666d992d