nova-manage map_instances is not using the cells info from the API database
In order to map instances, nova-manage takes the database info of the cell from
the config file (so by default this points to cell0 in nova.conf) and so for
every new cell created, in order to do the map_instances, the correct --config-file
has to be specified in addition to the --cell_uuid option that is provided.

So it is better if this info is taken from the cell_mappings table inside the
nova-api database using the provided --cell_uuid.

Change-Id: Ib86e25bf8c4cd4008eae43c49c74f24b1b63518a
Closes-Bug: #1720130