Make conductor pass and use host_lists
The earlier patches in the series generated alternates and Selection
objects, and modified the RPC calls to send them to the conductor. This
patch has the conductor pass these host_lists to the compute for the
build process, and, if the build fails, has the compute pass the
host_list back to the conductor.

Also fixes a bug in the scheduler manager exposed by this change
when using the CachingScheduler in a reschedule functional test.

Blueprint: return-alternate-hosts

Change-Id: Iae904afb6cb4fcea8bb27741d774ffbe986a5fb4