Split get_allocations_for_instance() into useful bits
We now call this method for other types of consumers, like migrations,
so the name needs to change. Also, we need to be able to fetch allocations
without specifying a resource_provider. This splits and renames the method
so it can be used in this way.

Related to blueprint migration-allocations

Change-Id: I2d27ae23aace1b7f14b5556af1693169391d4b2c