[placement] add coverage for update of standard resource class
In microversion <= 1.6 it is possible to update a resource class to have
a new name. This is only possible for custom resource classes, but there
was no test to show that the proper 400 was returned when attempting to
change a standard class.

This adds one test to cover it, and in the process changes the name of
a related test to distinguish it from the new one.

Change-Id: I6c665db68de91c0de91c16f7e622f0c7e2f28bc2
Closes-Bug: #1723123