[placement] Confirm that empty resources query causes 400
When a request is made to /resource_providers or /allocation_candidates
with a 'resources' query parameter, and empty value should lead to a
400. This was indeed happening, but there was no gabbi test coverage for
it. This adds tests at both URLs.

Note that because the query parameter is processed by the same code for
both /resource_providers and /allocation_candidates the test duplicates
a code a path, however, that's an implementation detail that we don't
know at the HTTP API layer.

Change-Id: Ia6fc0ac09efc5407a899de4215a2ba0ba0b017fb
Closes-Bug: #1723122