Regenerate context during targeting
We are only doing a copy.copy() on context when we do target_cell(),
which may mean we're sharing more across threads than we intend to.
This makes that a full regeneration of the context (like we would
do over RPC). This will necessarily dump the targeting, if already
set, and any per-thread database context.

Closes-Bug: #1722404
Change-Id: Id24dea7465bafc1f6f58c4a121c4ffb35b7a634a