XenAPI: get vGPU stats from hypervisor
This commit is to get vGPU resource stats from XenAPI.
It contains the following data for vGPU types:
* total amount:
  The capacity of the vGPUs which will be used to create
  inventory data for ResourceClass.VGPU.
* max_heads:
  The maximal display heads which will be used to create
  inventory data for ResourceClass.VGPU_DISPLAY_HEAD.
* model_name:
  The vGPU type's name in XenAPI. The vGPU types are filtered
  basing on model_name.
* PGPU group uuid and vGPU type's uuid:
  The identifiers in XenAPI. We will use them to create
  the child resource provider's name in nested-RP, so that
  the vGPU resource allocation can be easily mapped back
  to the PGPU group and vGPU type at spawning instance.

Once it's ready to use the nested resource providers, we
can support multiple vGPU types in a compute node. But
at the moment we only enable one vGPU type in a compute
node. The restriction is implemented by checking the
configure option of enabled_vgpu_type. If multiple vGPU
types are enabled, we will only use the first one.

Change-Id: I2f9cf80b71b50a4b8e011a4a8e40474cc02baba7
blueprint: add-support-for-vgpu