Pick ironic nodes without VCPU set
Starting with the Pike release, reporting VCPU/memory/disk is no longer required.
However, we used VCPU to check if a node is available, so nodes without VCPU in
their properties were always ignored. This patch changes the logic to use the existing
_node_resources_unavailable call.

This change also fixes another related issue: when disk or memory are missing from
properties, the virt driver tries to report zero max_unit for them, which is not
allowed by placement.

Change-Id: I1bbfc152189252c5c45e6153695a802d17b76690
Closes-Bug: #1723423