Send Allocations to spawn
This change adds a new 'allocations' kwarg to the
nova.virt.driver.ComputeDriver.spawn method and its overrides.  The
value is the 'allocations' member of the dict returned from GET
/allocations/{instance_uuid}, representing the resources allocated to
the instance from the scheduler via placement.

This provides the virt driver with a way to know exactly which inventory
has been allocated to an instance by placement, from which resource
providers.  Previously, virt drivers would process the flavor to glean
this information.  This works fine as long as there's one monolithic
resource provider - the compute host - providing all the inventory in
a known fashion.  However, as we expand to using nested and shared
resource providers, it becomes necessary for the virt driver to know
that e.g. CPU inventory has been allocated from a specific NUMA node; or
e.g. an SR-IOV VF has been allocated from a specific PF.  Placement
allows such allocations to be represented generically; this change
is how virt is made aware of the same.

Change-Id: I00eab47edf1150788777300680e853a872c1db40