Move restart_compute_service to a common place
Restarting a compute service properly in the functional test environment
is tricky. During a recent bugfix a util function was introduced to do
the restart. This patch moves the util to the base class and spreads
the usage of it to more functional tests to make them more realistic.

Change-Id: I17f67a02b27a90658df48856963ea3fb327e81dc
(cherry picked from commit 63eb7fee1bf95a2e09b120358d79445893633270)