libvirt: add check for VIR_DOMAIN_BLOCK_REBASE_COPY_DEV
Change I868a0dae0baf8cded9c7c5807ea63ffc5eec0c5e adds some
code that relies on the VIR_DOMAIN_BLOCK_REBASE_COPY_DEV flag
which was introduced to libvirt in commit
b7e73585a8d96677695a52bafb156f26cbd48fb5 which went into the
1.2.9 release:

https://libvirt.org/git/?p=libvirt.git;a=blob;f=docs/news-2014.html.in

The nova change went into the Pike release and was backported
to stable/ocata, and went into the 15.0.6 release. However,
in ocata the minimum required version of libvirt is 1.2.1, so
it's possible people could be running libvirt in ocata with
libvirt < 1.2.9 which would then result in an AttributeError
if doing a swap volume operation on a block type device.

This adds a check to see if the attribute exists before attempting
to use it.

Note that this change is not made on the master branch since starting
in Pike the minimum required version of libvirt is 1.2.9.

Change-Id: Ie3789bda6c6417249e279739b80a61c5b525421f
Closes-Bug: #1724039