Downgrade log for keystone verify client fail
Under some circumstances the keystone verify process might fail
but we are able to proceed because it's client setting error,
so we don't need to report an exception log in the log file to
confuse admin, instead, use an info log.

In the reported bug, the issue is that nova is configured for
the 'internal' identity endpoint but the nova code does not
pass an interface, so KSA defaults to 'public' which fails.
This is fixed with I2204c8bed8936d5bed0f410284d2a563f84e7100
but not something we can backport, so this is a simple change
to make the logging less annoying.

Closes-Bug: 1716344

Change-Id: I67c9f648f85de364de443e2a0535ddd361c14661
(cherry picked from commit 1606467b29969eb45efbb56c1b148a4a6f53c5cf)