Add test so we remember why CUSTOM_ prefix added
While it is tempting as an ironic operator to have CUSTOM_FOO in
placement, nova flavors and ironic resource class attributes,
allow that makes for a more complicated mapping between ironic
resource classes and the resource classes in the general case,
and could cause confusion when there is a typo of CUSTOM vs CUSTM
that could get mapped very differently into placement.

Adding this test so it is clear this is a deliberate choice, and
git blame will find this git commit message.

Related-Bug: #1724524

Change-Id: Ie38baea53e41ca7ba8875fd113cd777810f17db7